WIP anime episode torrent

+++ src/main/kotlin/fr/zakaoai/coldlibrarybackend/infrastructure/db/entities/AnimeEpisodeTorrent.kt
@@ -13,7 +13,7 @@ class AnimeEpisodeTorrent(
         val title: String,
         val date: LocalDate,
         val torrentLink: String,
-        val torrentId: Long,
+        val torrentId: Int,
         val torrentSize: String,
         val seeders: Int,
         val leechers: Int,
